Wrote CSR parser and logic to extract instance metadata from AWS
This initial commit includes functionality for reading the CSR file
supplied by the Puppet CA/Master, extracting out the CSR extensions
and looking up the instance ID against the AWS API.

It does not yet include the actual validation of the CSR extensions or
any sophisticated fault handling. It has also yet to be packaged as a
proper Gem.