Removed unique name validation.
Whilst well intentioned, this won't protect us from an attacker
re-sending an existing instance id. Our only solution here is that we
must enforce the unique aspect of instance IDs and track all instance
IDs that we have already signed. If a repeat comes in, we need to refuse
the signing request.